Added test coverage to the active editor controller.

+++ grails-app/domain/ocm/ActiveEditor.groovy
@@ -9,7 +9,7 @@ class ActiveEditor {
   // Using a string for the mapId instead a foreign key relation since we only
   // need to use the id as way to aggregate the common editors.
   String mapId
-  float lastCheckIn = System.currentTimeMillis()
+  long lastCheckIn = System.currentTimeMillis()
 
   static constraints = {
   }
